Added missing check whether incoming entity actually exists. Without check, the backend database sequence gets incremented by declared step.

+++ src/main/java/gr/codehub/guide/filmrepository/controller/AbstractDomainController.java
@@ -50,7 +50,9 @@ public abstract class AbstractDomainController<T> {
 
 	@DeleteMapping
 	public void delete(@Valid @RequestBody final T entity) {
-		getDomainService().delete(entity);
+		if (getDomainService().exists(entity)) {
+			getDomainService().delete(entity);
+		}
 	}
 
 	@PatchMapping("/{id}")
